==> 03_ABCD meas paper mixed models.Rmd <==
---
title: "ABCD measurement paper mixed models"
author: "Meriah DeJoseph (edited by R. Sifre)"
date: "2020 - 2021"
output:
  html_document:
    toc: true
    toc_float: true
---

```{r message=FALSE}
library(tidyverse)
# Multi-level modeling 
library(lme4)
library(lmerTest)
library(bbmle)
library(AICcmodavg)
library(interactions)
library(huxtable)
library(ggstance)

# Imputation 
library(mice)
library(miceadds)
library(finalfit)

# Train/test split
library(caret)

library(knitr)

knitr::opts_chunk$set(echo = TRUE)
```


# About
This file contains the code for the statistical models for this paper. It assumes that there are 3 .csvs called `imputed_dat.csv`, `train_dat.csv` , and `test_dat.csv` in your working directory (generated by `ABCD-meas-paper-impute-split-data.Rmd` ).  

## Overall steps:
1. - Fit models on the training set. Keep sig ones when testing functional form for vars determined may be quadratic in prelim analyses  
2. - Test models. Run the final chosen models on the test data  
3. - For models that performed above chance (Rsq>0), run model on full dataset and get parameters to report in ms 
4. - Plot results on full dataset (as long as model performed above chance)  

# Read data
```{r}
# Set working directory
setwd("~/Box/!ICD/ABCD/ANALYSES/!REVISION") #meriah's
#setwd("~/Box/ANALYSES") #robin's

train = read.csv(paste(getwd(), 'train_dat.csv', sep='/'))
test = read.csv(paste(getwd(), 'test_dat.csv', sep='/'))
data_1 = read.csv(paste(getwd(), 'imputed_dat.csv', sep='/'))
```


# Fit Models 

Model fitting steps:  
1. - Fit with all random and fixed main effects including motion  
2a. - Fit quadratic effects for those deemed they may be quadratic and eliminate  
ones that aren't sig  
2b. - If applicable, compare quad model (step 2) vs nonquad model (step 1) (figure out model comparison cutoff)  
3. - Fit interactions of interest (inc w/ quad vars retained in step 3)  
4. - If applicable, compare interaction model with one before and choose the best based on AIC plus sig log likelihood  
5. - Make predictions using test data to assess whether model performed above chance in test set. If yes (Rsq>0), then plot the final model and report results using full dataset  

### Set up equations
```{r echo=TRUE}
# Baseline vars (all fixed main effects that must be included bc of MNLFA requirements, and random effects of intercept)
equation_1 = '1+ rsfmri_c_ngd_meanmotion + AGEC + FEMALE + BLACK + HISPANIC + OTHER + INRC + HIGHEDC + DepETA + ThreatETA + SocETA + (1|sitenum/rel_family_id)'

# Equation 1 + Income interactions 
equation_1_plus_INRinteractions = '1+ rsfmri_c_ngd_meanmotion + AGEC + FEMALE + BLACK + HISPANIC + OTHER + INRC + HIGHEDC + DepETA + ThreatETA + SocETA + INRC:ThreatETA + INRC:SocETA + (1|sitenum/rel_family_id)'

# Equation 1 + Ed interactions 
equation_1_plus_EDinteractions = '1+ rsfmri_c_ngd_meanmotion + AGEC + FEMALE + BLACK + HISPANIC + OTHER + INRC + HIGHEDC + DepETA + ThreatETA + SocETA + HIGHEDC:ThreatETA + HIGHEDC:SocETA + (1|sitenum/rel_family_id)'
```

### Function for calculating model fit on held out data
```{r}
getR2 <- function(mod, outcome, test_dat, train_dat) {
  #Use final model selected in w/ train data above and test on holdout test data
  yhattest <- predict(mod, newdata = test_dat, 
                      re.form = NA) # Set to NA bc held-out data has different groupings
  
  #### calculate r2 ####
  # calculate SSE
  #resids <- test$conWith - yhattest #residuals
  resids <- test[, outcome] - yhattest #residuals
  
  SSE <- sum(resids^2) #estimate of how bad this model is

  # calculate SST
  deviations <- test_dat[, outcome] - mean(train_dat[, outcome]) #null model
  SST <- sum(deviations^2)

  # calculate r2
  R2 <- 1 - SSE/SST
  
  return(R2)

}
```



# CON (within-net FC)

## Fit models

1. Fit with all random and fixed main effects including motion
```{r conWith_fit1, echo=TRUE, include=TRUE, cache=TRUE, message=FALSE}
conWithlmer1 <- lmer(paste('conWith', equation_1, sep = '~'), 
                    data = train, REML = FALSE,
                    control=lmerControl(optimizer = "Nelder_Mead"))
summary(conWithlmer1) 
```

2. Fit quadratic effects
```{r conWith_fit2, echo=TRUE, include=TRUE, cache=TRUE, message=FALSE}
equation_2 = paste(equation_1, '+ DepETA_2 + SocETA_2 + INRC_2 ')
conWithlmer2 <- lmer(paste('conWith', equation_2, sep = '~'), 
                    data = train, REML = FALSE,
                    control=lmerControl(optimizer = "Nelder_Mead"))
summary(conWithlmer2)
```
No significant quadratic effects

3. Fit interactions  
```{r conWith_fit3, echo=TRUE, include=TRUE, cache=TRUE, message=FALSE}
#Income
conWithlmer3 <- lmer(paste('conWith', equation_1_plus_INRinteractions, sep = '~'), 
                     data = train, REML = FALSE)
summary(conWithlmer3) #nothing, so moving on to edu now

# Edu
conWithlmer4 <- lmer(paste('conWith', equation_1_plus_EDinteractions, sep = '~'), 
                     data = train, REML = FALSE)
summary(conWithlmer4) #none so going with model 1
```
Final model = Model 1.

## Test robustness with held-out data
```{r conWithTest, cache=TRUE, echo=TRUE}
R2 <- getR2(mod=conWithlmer1, outcome='conWith', test_dat=test, train_dat=train) 
R2 #0.05512193 REVISION:0.05572414
```

## Final model with complete dataset
```{r conWithFinal, warning=FALSE, echo=TRUE, include=TRUE, cache=TRUE}
conWithFINAL <- lmer(paste('conWith', equation_1, sep = '~'), 
                    data = data_1, REML = FALSE,
                    control=lmerControl(optimizer = "Nelder_Mead"))
summary(conWithFINAL) #INR sig REVISION: INR no longer sig, edu and depETA trend
```


# DMN (within-net FC)

## Fit models

1. Fit with all random and fixed main effects including motion
```{r dmnWith_fit1, echo=TRUE, include=TRUE, cache=TRUE, message=FALSE}
dmnWithlmer1 <- lmer(paste('dmnWith', equation_1, sep = '~'), 
                    data = train, REML = FALSE,
                    control=lmerControl(optimizer = "Nelder_Mead"))
summary(dmnWithlmer1) 
```

2. Fit quadratic effects
```{r dmnWith_fit2, echo=TRUE, include=TRUE, cache=TRUE, message=FALSE}
equation_2 = paste(equation_1, '+ DepETA_2 + SocETA_2 + INRC_2 ')
dmnWithlmer2 <- lmer(paste('dmnWith', equation_2, sep = '~'), 
                    data = train, REML = FALSE,
                    control=lmerControl(optimizer = "Nelder_Mead"))
summary(dmnWithlmer2)
```
No significant quadratic effects  


3. Fit interactions  
```{r dmnWith_fit3, echo=TRUE, include=TRUE, cache=TRUE, message=FALSE}
#Income
dmnWithlmer3 <- lmer(paste('dmnWith', equation_1_plus_INRinteractions, sep = '~'), 
                     data = train, REML = FALSE)
summary(dmnWithlmer3) #nothing, so moving on to edu now

# Edu
dmnWithlmer4 <- lmer(paste('dmnWith', equation_1_plus_EDinteractions, sep = '~'), 
                     data = train, REML = FALSE)
summary(dmnWithlmer4) #none so going with model 1
```
Final model = Model 1.

## Test robustness with held-out data
```{r dmnWithTest, cache=TRUE, echo=TRUE}
R2 <- getR2(mod=dmnWithlmer1, outcome='dmnWith', test_dat=test, train_dat=train) 
R2 #0.03883489 REVISION: 0.03960668
```

## Final model with complete dataset
```{r dmnWithFinal, warning=FALSE, echo=TRUE, include=TRUE, cache=TRUE}
dmnWithFINAL <- lmer(paste('dmnWith', equation_1, sep = '~'), 
                    data = data_1, REML = FALSE,
                    control=lmerControl(optimizer = "Nelder_Mead"))

summary(dmnWithFINAL) #INR trend; REVISION: INR and SocETA sig

```

## Plot effects -- new w/ revision
```{r conWithFinalInfo, warning=FALSE, echo=TRUE, include=TRUE, cache=TRUE}
fdmnWith=fixef(dmnWithFINAL)
fdmnWith["INRC"]*sd(data_1$INRC, na.rm=TRUE)/sd(data_1$dmnWith) #0.03915204 
fdmnWith["SocETA"]*sd(data_1$SocETA, na.rm=TRUE)/sd(data_1$dmnWith) #0.02251698 

#Get predicted values
data_1$dmnWithyhat = predict(dmnWithFINAL, newdata = data_1, re.form = NA)

#Plot sig main effects of interest
#First get +/- 2 SD for y axis
mean(data_1$dmnWith) + 2*(sd(data_1$dmnWith)) #0.4033402
mean(data_1$dmnWith) - 2*(sd(data_1$dmnWith)) #0.09807931

dmnWithlmerINR <- ggplot(data = data_1, aes(x = INR, y = dmnWithyhat)) +
  geom_jitter(aes(y=dmnWith),alpha = 0.1, size=.7) +
  geom_smooth(method="lm", level=.95, lwd = 2, color = 'blue') +
  theme_classic() +
  theme(axis.title.x = element_text(size=15),
        axis.title.y = element_text(size=15),
        axis.text.x = element_text (size=10),
        axis.text.y = element_text (size=10)) +
  scale_y_continuous(limits = c(0.09807931, 0.4033402)) +
  ylab("Default Mode \n Network Connectivity") +
  xlab("Family Income-to-Needs Ratio (INR)")
dmnWithlmerINR
ggsave("dmnWithlmerINR.png", width = 5, height = 4)

dmnWithlmerSOC <- ggplot(data = data_1, aes(x = SocETA, y = dmnWithyhat)) +
  geom_jitter(aes(y=dmnWith),alpha = 0.1, size=.7) +
  geom_smooth(method="lm", level=.95, lwd = 2, color = 'blue') +
  theme_classic() +
  theme(axis.title.x = element_text(size=15),
        axis.title.y = element_text(size=15),
        axis.text.x = element_text (size=10),
        axis.text.y = element_text (size=10)) +
  scale_y_continuous(limits = c(0.09807931, 0.4033402)) +
  ylab("Default Mode \n Network Connectivity") +
  xlab("Caregiver Social Support (MNLFA Score)")
dmnWithlmerSOC
ggsave("dmnWithlmerSOC.png", width = 5, height = 4)

```



# DAN (within-net) 

## Fit models

1. Fit with all random and fixed main effects including motion
```{r danWith_fit1, echo=TRUE, include=TRUE, cache=TRUE, message=FALSE}
danWithlmer1 <- lmer(paste('danWith', equation_1, sep = '~'), 
                    data = train, REML = FALSE,
                    control=lmerControl(optimizer = "Nelder_Mead"))
summary(danWithlmer1) 
```

2. Fit quadratic effects
```{r danWith_fit2, echo=TRUE, include=TRUE, cache=TRUE, message=FALSE}
equation_2 = paste(equation_1, '+DepETA_2 + INRC_2') 
danWithlmer2 <- lmer(paste('danWith', equation_2, sep = '~'), 
                    data = train, REML = FALSE,
                    control=lmerControl(optimizer = "Nelder_Mead"))
summary(danWithlmer2)
```
No significant quadratic effects

3. Fit interactions  
```{r danWith_fit3, echo=TRUE, include=TRUE, cache=TRUE, message=FALSE}
#Income
danWithlmer3 <- lmer(paste('danWith', equation_1_plus_INRinteractions, sep = '~'), 
                     data = train, REML = FALSE)
summary(danWithlmer3) #nothing, so moving on to edu now

# Edu
danWithlmer4 <- lmer(paste('danWith', equation_1_plus_EDinteractions, sep = '~'), 
                     data = train, REML = FALSE)
summary(danWithlmer4) #none so going with model 1
```
Final model = Model 1.

## Test robustness with held-out data
```{r danWithTest, cache=TRUE, echo=TRUE}
R2 <- getR2(mod=danWithlmer1, outcome='danWith', test_dat=test, train_dat=train) 
R2 #0.01798139 REVISION: 0.01886793
```

## Final model with complete dataset
```{r danWithFinal, warning=FALSE, echo=TRUE, include=TRUE, cache=TRUE}
danWithFINAL <- lmer(paste('danWith', equation_1, sep = '~'), 
                    data = data_1, REML = FALSE,
                    control=lmerControl(optimizer = "Nelder_Mead"))
summary(danWithFINAL) #highed, INR trend REVISION: highed trend only
```

# VAN (within-net) 
## Fit models

1. Fit with all random and fixed main effects including motion
```{r vanWith_fit1, echo=TRUE, include=TRUE, cache=TRUE, message=FALSE}
vanWithlmer1 <- lmer(paste('vanWith', equation_1, sep = '~'), 
                    data = train, REML = FALSE,
                    control=lmerControl(optimizer = "Nelder_Mead"))
summary(vanWithlmer1) 
```

2. Fit quadratic effects
```{r vanWith_fit2, echo=TRUE, include=TRUE, cache=TRUE, message=FALSE}
equation_2 = paste(equation_1, '+ INRC_2')
vanWithlmer2 <- lmer(paste('vanWith', equation_2, sep = '~'), 
                    data = train, REML = FALSE,
                    control=lmerControl(optimizer = "Nelder_Mead"))
summary(vanWithlmer2)
```
No sig. quadratic effects

3. Fit interactions  
```{r vanWith_fit3, echo=TRUE, include=TRUE, cache=TRUE, message=FALSE}
#Income
vanWithlmer3 <- lmer(paste('vanWith', equation_1_plus_INRinteractions, sep = '~'), 
                     data = train, REML = FALSE)
summary(vanWithlmer3) #nothing, so moving on to edu now

# Edu
vanWithlmer4 <- lmer(paste('vanWith', equation_1_plus_EDinteractions, sep = '~'), 
                     data = train, REML = FALSE)
summary(vanWithlmer4) #none so going with model 1
```
Final model = Model 1.

## Test robustness with held-out data
```{r vanWithTest, cache=TRUE, echo=TRUE}
R2 <- getR2(mod=vanWithlmer1, outcome='vanWith', test_dat=test, train_dat=train) 
R2 #0.006768083 REVISION: 0.007856234
```

## Final model with complete dataset
```{r vanWithFinal, warning=FALSE, echo=TRUE, include=TRUE, cache=TRUE}
vanWithFINAL <- lmer(paste('vanWith', equation_1, sep = '~'), 
                    data = data_1, REML = FALSE,
                    control=lmerControl(optimizer = "Nelder_Mead"))
summary(vanWithFINAL) # INR REVISION: DepETA trend only
```


# fPAR (within-net)
## Fit models

1. Fit with all random and fixed main effects including motion
```{r fparWith_fit1, echo=TRUE, include=TRUE, cache=TRUE, message=FALSE}
fparWithlmer1 <- lmer(paste('fparWith', equation_1, sep = '~'), 
                    data = train, REML = FALSE,
                    control=lmerControl(optimizer = "Nelder_Mead"))
summary(fparWithlmer1) 
```

2. Did not see quadratic trend in preliminary analyses - do not test for this in models 

3. Fit interactions  
```{r fparWith_fit3, echo=TRUE, include=TRUE, cache=TRUE, message=FALSE}
#Income
fparWithlmer3 <- lmer(paste('fparWith', equation_1_plus_INRinteractions, sep = '~'), 
                     data = train, REML = FALSE)
summary(fparWithlmer3) #nothing, so moving on to edu now

# Edu
fparWithlmer4 <- lmer(paste('fparWith', equation_1_plus_EDinteractions, sep = '~'), 
                     data = train, REML = FALSE)
summary(fparWithlmer4) #highedxthreat sig REVISION: now only trend

# equation_1_plus_Edthreat = paste(equation_1, '+ HIGHEDC:ThreatETA')
# 
# fparWithlmer4 <- lmer(paste('fparWith', equation_1_plus_Edthreat, sep = '~'), 
#                      data = train, REML = FALSE)
# summary(fparWithlmer4)
# 
# anova(fparWithlmer1, fparWithlmer4) #model 4 best
```
Final model = Model 1.


## Test robustness with held-out data
```{r fparWithTest, cache=TRUE, echo=TRUE}
R2 <- getR2(mod=fparWithlmer1, outcome='fparWith', test_dat=test, train_dat=train) 
R2 #0.01775183 REVISION: 0.01729793
```

## Final model with complete dataset
```{r fparWithFinal, warning=FALSE, echo=TRUE, include=TRUE, cache=TRUE}
fparWithFINAL <- lmer(paste('fparWith', equation_1, sep = '~'), 
                    data = data_1, REML = FALSE,
                    control=lmerControl(optimizer = "Nelder_Mead"))
summary(fparWithFINAL) #edu x threat sig REVISION: nothing sig
```


# Sal (within-net) 
## Fit models

1. Fit with all random and fixed main effects including motion
```{r salWith_fit1, echo=TRUE, include=TRUE, cache=TRUE, message=FALSE}
salWithlmer1 <- lmer(paste('salWith', equation_1, sep = '~'), 
                    data = train, REML = FALSE,
                    control=lmerControl(optimizer = "Nelder_Mead"))
summary(salWithlmer1) 
```

2. Fit quadratic effects
```{r}
equation_2 = paste(equation_1, '+ INRC_2')
salWithlmer2 <- lmer(paste('salWith', equation_2, sep = '~'), 
                    data = train, REML = FALSE,
                    control=lmerControl(optimizer = "Nelder_Mead"))
summary(salWithlmer2) 

```
No sig. quadratic effects

3. Fit interactions  
```{r salWith_fit3, echo=TRUE, include=TRUE, cache=TRUE, message=FALSE}
#Income
salWithlmer3 <- lmer(paste('salWith', equation_1_plus_INRinteractions, sep = '~'), 
                     data = train, REML = FALSE)
summary(salWithlmer3) #nothing, so moving on to edu now

# Edu
salWithlmer4 <- lmer(paste('salWith', equation_1_plus_EDinteractions, sep = '~'), 
                     data = train, REML = FALSE)
summary(salWithlmer4) #none so going with model 1
```
Final model = Model 1.

## Test robustness with held-out data
```{r salWithTest, cache=TRUE, echo=TRUE}
R2 <- getR2(mod=salWithlmer1, outcome='salWith', test_dat=test, train_dat=train) 
R2 #0.005385729 REVISION: 0.006647365
```

## Final model with complete dataset
```{r salWithFinal, warning=FALSE, echo=TRUE, include=TRUE, cache=TRUE}
salWithFINAL <- lmer(paste('salWith', equation_1, sep = '~'), 
                    data = data_1, REML = FALSE,
                    control=lmerControl(optimizer = "Nelder_Mead"))
summary(salWithFINAL) #HIGHED REVISION: still highed w. soceta trend
```


Examine effect sizes of significant predictors, plot final mod 
```{r, warning=FALSE, echo=TRUE, include=TRUE, cache=TRUE}
#Examine effect sizes of sig predictors
fsalWith=fixef(salWithFINAL)
fsalWith["HIGHEDC"]*sd(data_1$HIGHEDC, na.rm=TRUE)/sd(data_1$salWith) #0.04533088 REVISION: 0.03633923  

#Get predicted values
data_1$salWithyhat = predict(salWithFINAL, newdata = data_1, re.form = NA)

#Plot sig main effects of interest 
#First get +/- 2 SD for y axis 
mean(data_1$salWith) + 2*(sd(data_1$salWith)) #0.7103859
mean(data_1$salWith) - 2*(sd(data_1$salWith)) #0.05866069

salWithlmerHIGHED<- ggplot(data = data_1, aes(x = HIGHED, y = salWithyhat)) +
  geom_jitter(aes(y=salWith),alpha = 0.1, size = .7) + 
  stat_smooth(method="lm", level = .95, lwd = 2, show.legend = FALSE) +
  theme_classic() +
   theme(axis.title.x = element_text(size=15),
        axis.title.y = element_text(size=15),
        axis.text.x = element_text (size=10),
        axis.text.y = element_text (size=10)) +
  scale_y_continuous(limits = c(0.05866069, 0.7103859))+
  ylab("Salience Network Connectivity") +
  xlab("Highest Parental Education (Years)")
salWithlmerHIGHED
ggsave("salWithlmerHIGHED.png", width = 5, height = 4)
```


# Sal-Dmn (Between-net) models
## Fit models

1. Fit with all random and fixed main effects including motion
```{r salDmn_fit1, echo=TRUE, include=TRUE, cache=TRUE, message=FALSE}
salDmnlmer1 <- lmer(paste('salDmn', equation_1, sep = '~'), 
                    data = train, REML = FALSE,
                    control=lmerControl(optimizer = "Nelder_Mead"))
summary(salDmnlmer1) 
```

2. Nothing looked quadratic, skip step of fitting quadratic model


3. Fit interactions  
```{r salDmn_fit3, echo=TRUE, include=TRUE, cache=TRUE, message=FALSE}
#Income
salDmnlmer3 <- lmer(paste('salDmn', equation_1_plus_INRinteractions, sep = '~'), 
                     data = train, REML = FALSE)
summary(salDmnlmer3) #nothing, so moving on to edu now

# Edu
salDmnlmer4 <- lmer(paste('salDmn', equation_1_plus_EDinteractions, sep = '~'), 
                     data = train, REML = FALSE)
summary(salDmnlmer4) #none so going with model 1
```
Final model = Model 1.

## Test robustness with held-out data
```{r salDmnTest, cache=TRUE, echo=TRUE}
R2 <- getR2(mod=salDmnlmer1, outcome='salDmn', test_dat=test, train_dat=train) 
R2 #0.01346418 REVISION: 0.01291619
```

## Final model with complete dataset
```{r salDmnFinal, warning=FALSE, echo=TRUE, include=TRUE, cache=TRUE}
salDmnFINAL <- lmer(paste('salDmn', equation_1, sep = '~'), 
                    data = data_1, REML = FALSE,
                    control=lmerControl(optimizer = "Nelder_Mead"))
summary(salDmnFINAL) #nothing of interest REVISION: soceta trend
```

# danVan (Between-Net)
## Fit models

1. Fit with all random and fixed main effects including motion
```{r danVan_fit1, echo=TRUE, include=TRUE, cache=TRUE, message=FALSE}
danVanlmer1 <- lmer(paste('danVan', equation_1, sep = '~'), 
                    data = train, REML = FALSE,
                    control=lmerControl(optimizer = "Nelder_Mead"))
summary(danVanlmer1) 
```

2.Fit quadratic effects
```{r danVan_fit2, echo=TRUE, include=TRUE, cache=TRUE, message=FALSE}
equation_2 = paste(equation_1, '+ SocETA_2')
danVanlmer2 <- lmer(paste('danVan', equation_2, sep = '~'), 
                    data = train, REML = FALSE,
                    control=lmerControl(optimizer = "Nelder_Mead"))
summary(danVanlmer2) 
```
No sig. quadratic effects


3. Fit interactions  
```{r danVan_fit3, echo=TRUE, include=TRUE, cache=TRUE, message=FALSE}
#Income
danVanlmer3 <- lmer(paste('danVan', equation_1_plus_INRinteractions, sep = '~'), 
                     data = train, REML = FALSE)
summary(danVanlmer3) #INR x threat int

equation_1_plus_INRthreat = paste(equation_1, '+ INRC:ThreatETA ')
danVanlmer3 <- lmer(paste('danVan', equation_1_plus_INRthreat, sep = '~'), 
                     data = train, REML = FALSE)
summary(danVanlmer3)

# Edu
danVanlmer4 <- lmer(paste('danVan', equation_1_plus_EDinteractions, sep = '~'), 
                     data = train, REML = FALSE)
summary(danVanlmer4) #none 

#compare
anova(danVanlmer1, danVanlmer3) #model 3 better
```
Final model = Model 3.

## Test robustness with held-out data
```{r danVanTest, cache=TRUE, echo=TRUE}
R2 <- getR2(mod=danVanlmer3, outcome='danVan', test_dat=test, train_dat=train) 
R2 #0.02603877 REVISION: 0.02761846
```

## Final model with complete dataset
```{r danVanFinal, warning=FALSE, echo=TRUE, include=TRUE, cache=TRUE}
danVanFINAL <- lmer(paste('danVan', equation_1_plus_INRthreat, sep = '~'), 
                    data = data_1, REML = FALSE,
                    control=lmerControl(optimizer = "Nelder_Mead"))
summary(danVanFINAL) #INR, threat, soc, inr x threat REVISION: Threat and INRxthreat sig w/ soceta trend
```

## Examine effect sizes of significant predictors, plot final mod 
```{r, warning=FALSE, echo=TRUE, include=TRUE, cache=TRUE}
fdanVan=fixef(danVanFINAL)
#fdanVan["INRC"]*sd(data_1$INRC, na.rm=TRUE)/sd(data_1$danVan) #-0.03431884 
fdanVan["ThreatETA"]*sd(data_1$ThreatETA, na.rm=TRUE)/sd(data_1$danVan) #0.03293605 REVISION: 0.03317738
fdanVan["SocETA"]*sd(data_1$SocETA, na.rm=TRUE)/sd(data_1$danVan) #0.03021621 REVISION: 0.02236433 

#Get predicted values
data_1$danVanyhat = predict(danVanFINAL, newdata = data_1, re.form = NA)

#Plot sig main effects of interest 
#First get +/- 2 SD for y axis 
mean(data_1$danVan) + 2*(sd(data_1$danVan)) #0.03750963
mean(data_1$danVan) - 2*(sd(data_1$danVan)) #-0.2068818

danVanlmerSOC <- ggplot(data = data_1, aes(x = SocETA, y = danVanyhat)) +
  geom_jitter(aes(y=danVan),alpha = 0.1, size=.7) + 
  stat_smooth(method="lm", level = .95, lwd = 2, show.legend = FALSE) +
  theme_classic() +
  theme(axis.title.x = element_text(size=13.7),
        axis.title.y = element_text(size=15),
        axis.text.x = element_text (size=10),
        axis.text.y = element_text (size=10)) +
 scale_y_continuous(limits = c(-0.2068818, 0.03750963)) +
  ylab("Dorsal and Ventral Attention \n Between-Network Connectivity") +
  xlab("Caregiver Social Support (MNLFA Score)")
danVanlmerSOC
ggsave("danVanlmerSOC.png", width = 5, height = 4)


##Plot interaction##

#median splits 
danVanlmerINRxTHREAT <- ggplot(data = data_1, aes(x = INR, y = danVanyhat, color= as.factor(ThreatETAmed))) +
  stat_smooth(method="lm", level = .95, lwd = 2, show.legend = FALSE) +
  theme_classic() +
  theme(axis.title.x = element_text(size=15),
        axis.title.y = element_text(size=15),
        axis.text.x = element_text (size=10),
        axis.text.y = element_text (size=10)) +
  ylab("Dorsal and Ventral Attention \n Between-Network Connectivity") +
  xlab("Family Income-to-Needs Ratio (INR)")
danVanlmerINRxTHREAT
ggsave("danVanlmerINRxTHREAT.png", width = 5, height = 4)


#Plot now with simple slopes
ss=sim_slopes(model=danVanFINAL, pred=INRC, modx=ThreatETA, jnplot=T, confint = T, round=4)
ss
plot(ss)
as_huxtable(ss)
#Johnson-Neyman plot shows that both the 95 CI lines are below 0, then the link between INR and conAmy is sig. So, with high values of threat, it is sig. The reverse is not true. So for those with low threat, the link between INR and comAmy is not sig.


FINALdanVanlmerINRxTHREAT <- interact_plot(danVanFINAL, pred = INRC, modx = ThreatETA, interval = T, legend.main = "Psychosocial Threat \n (MNLFA Score)", colors = "Paired", line.thickness = 2)  +
  theme_classic() + 
  theme(axis.title.x = element_text(size=20),
        axis.title.y = element_text(size=20),
        axis.text.x = element_text (size=15),
        axis.text.y = element_text (size=15),
        legend.text = element_text (size=15),
        legend.title = element_text (size=17),
        legend.key.width = unit(3.5, "line")) +
  ylab("Dorsal and Ventral Attention Between-Network Connectivity") +
  xlab("Family Income-to-Needs Ratio (INR)") 
FINALdanVanlmerINRxTHREAT 
ggsave("FINALdanVanlmerINRxTHREAT .png", width = 11, height = 8)

```




# conAmy_2h models
## Fit models

1. Fit with all random and fixed main effects including motion
```{r conAmy_2h_fit1, echo=TRUE, include=TRUE, cache=TRUE, message=FALSE}
conAmy_2hlmer1 <- lmer(paste('conAmy_2h', equation_1, sep = '~'), 
                    data = train, REML = FALSE,
                    control=lmerControl(optimizer = "Nelder_Mead"))
summary(conAmy_2hlmer1) 
```

2.Fit quadratic effects
```{r conAmy_2h_fit2, echo=TRUE, include=TRUE, cache=TRUE, message=FALSE}
equation_2a = paste(equation_1, '+ INRC_2 + HIGHEDC_2')
conAmy_2hlmer2a <- lmer(paste('conAmy_2h', equation_2a, sep = '~'), 
                    data = train, REML = FALSE,
                    control=lmerControl(optimizer = "Nelder_Mead"))
summary(conAmy_2hlmer2a) 
```
No sig. quadratic effects

3. Fit interactions  
```{r conAmy_2h_fit3, echo=TRUE, include=TRUE, cache=TRUE, message=FALSE}
#Income
conAmy_2hlmer3 <- lmer(paste('conAmy_2h', equation_1_plus_INRinteractions, sep = '~'), 
                     data = train, REML = FALSE)
summary(conAmy_2hlmer3) #nothing, so moving on to edu now

# Edu
conAmy_2hlmer4 <- lmer(paste('conAmy_2h', equation_1_plus_EDinteractions, sep = '~'), 
                     data = train, REML = FALSE)
summary(conAmy_2hlmer4) #none so going with model 1
```
Final model = Model 1.

## Test robustness with held-out data
```{r conAmy_2hTest, cache=TRUE, echo=TRUE}
R2 <- getR2(mod=conAmy_2hlmer1, outcome='conAmy_2h', test_dat=test, train_dat=train) 
R2 #0.03498296
```

## Final model with complete dataset
```{r conAmy_2hFinal, warning=FALSE, echo=TRUE, include=TRUE, cache=TRUE}
conAmy_2hFINAL <- lmer(paste('conAmy_2h', equation_1, sep = '~'), 
                    data = data_1, REML = FALSE,
                    control=lmerControl(optimizer = "Nelder_Mead"))
summary(conAmy_2hFINAL) #INR and threat main effect REVISION: just threat
```

##Examine effect sizes of significant predictors, plot final mod 
```{r, warning=FALSE, echo=TRUE, include=TRUE, cache=TRUE}
#Examine effect sizes of sig predictors
fconAmy_2h=fixef(conAmy_2hFINAL)
#fconAmy_2h["INRC"]*sd(data_1$INRC, na.rm=TRUE)/sd(data_1$conAmy_2h) #0.0366404
fconAmy_2h["ThreatETA"]*sd(data_1$ThreatETA, na.rm=TRUE)/sd(data_1$conAmy_2h) #-0.04176298 REVISION: -0.04517985 

#Get predicted values
data_1$conAmy_2hyhat = predict(conAmy_2hFINAL, newdata = data_1, re.form = NA)

#Plot sig main effects of interest 
#First get +/- 2 SD for y axis 
mean(data_1$conAmy_2h) + 2*(sd(data_1$conAmy_2h)) #0.2951599
mean(data_1$conAmy_2h) - 2*(sd(data_1$conAmy_2h)) #-0.2774557

# conAmy_2hlmerINR <- ggplot(data = data_1, aes(x = INR, y = conAmy_2hyhat)) +
#   geom_jitter(aes(y=conAmy_2h),alpha = 0.1, size=.7) +
#   stat_smooth(method="lm", level = .95, lwd = 2, show.legend = FALSE) +
#   theme_classic() +
#     theme(axis.title.x = element_text(size=15),
#         axis.title.y = element_text(size=15),
#         axis.text.x = element_text (size=10),
#         axis.text.y = element_text (size=10)) +
#   scale_y_continuous(limits = c(-0.2774557, 0.2951599)) +
#   ylab("Cingulo-Opercular and Amygdala \n Connectivity") +
#   xlab("Family Income-to-Needs Ratio (INR)")
# conAmy_2hlmerINR
# ggsave("conAmy_2hlmerINR.png", width = 5, height = 4)

conAmy_2hlmerTHREAT <- ggplot(data = data_1, aes(x = ThreatETA, y = conAmy_2hyhat)) +
  geom_jitter(aes(y=conAmy_2h),alpha = 0.1, size=.7) +
  stat_smooth(method="lm", level = .95, lwd = 2, show.legend = FALSE) +
  theme_classic() +
      theme(axis.title.x = element_text(size=15),
        axis.title.y = element_text(size=15),
        axis.text.x = element_text (size=10),
        axis.text.y = element_text (size=10)) +
  scale_y_continuous(limits = c(-0.2774557, 0.2951599)) +
  ylab("Cingulo-Opercular and Amygdala \n Connectivity") +
  xlab("Psychosocial Threat (MNLFA Score)")
conAmy_2hlmerTHREAT
ggsave("conAmy_2hlmerTHREAT.png", width = 5, height = 4)

```



# conHipp_2h models

## Fit models
1. Fit with all random and fixed main effects including motion
```{r conHipp_fit1, echo=TRUE, include=TRUE, cache=TRUE, message=FALSE}
conHipp_2hlmer1 <- lmer(paste('conHipp_2h', equation_1, sep = '~'), 
                    data = train, REML = FALSE,
                    control=lmerControl(optimizer = "Nelder_Mead"))
summary(conHipp_2hlmer1) 
```

2.Fit quadratic effects
```{r conHipp_2h_fit2, echo=TRUE, include=TRUE, cache=TRUE, message=FALSE}
equation_2a = paste(equation_1, '+ INRC_2 + HIGHEDC_2 + HIGHEDC_2')
conHipp_2hlmer2a <- lmer(paste('conHipp_2h', equation_2a, sep = '~'), 
                    data = train, REML = FALSE,
                    control=lmerControl(optimizer = "Nelder_Mead"))
summary(conHipp_2hlmer2a) 
```
No sig. quadratic effects

3. Fit interactions  
```{r conHipp_2h_fit3, echo=TRUE, include=TRUE, cache=TRUE, message=FALSE}
#Income
conHipp_2hlmer3 <- lmer(paste('conHipp_2h', equation_1_plus_INRinteractions, sep = '~'), 
                     data = train, REML = FALSE,
                     control=lmerControl(optimizer = "Nelder_Mead"))
summary(conHipp_2hlmer3) #nothing, so moving on to edu now

# Edu
conHipp_2hlmer4 <- lmer(paste('conHipp_2h', equation_1_plus_EDinteractions, sep = '~'), 
                     data = train, REML = FALSE,
                     control=lmerControl(optimizer = "Nelder_Mead"))
summary(conHipp_2hlmer4) #none so going with model 1
```
Final model = Model 1.

## Test robustness with held-out data
```{r conHipp_2hTest, cache=TRUE, echo=TRUE}
R2 <- getR2(mod=conHipp_2hlmer1, outcome='conHipp_2h', test_dat=test, train_dat=train) 
R2 #0.02508125 REVISION: 0.02660078
```

## Final model with complete dataset
```{r conHipp_2hFinal, warning=FALSE, echo=TRUE, include=TRUE, cache=TRUE}
conHipp_2hFINAL <- lmer(paste('conHipp_2h', equation_1, sep = '~'), 
                    data = data_1, REML = FALSE,
                    control=lmerControl(optimizer = "Nelder_Mead"))
summary(conHipp_2hFINAL) #highed sig REVISION: highed and threat sig
```

##Examine effect sizes of significant predictors, plot final mod 
```{r, warning=FALSE, echo=TRUE, include=TRUE, cache=TRUE}
#Examine effect sizes of sig predictors
fconHipp_2h=fixef(conHipp_2hFINAL)
fconHipp_2h["HIGHEDC"]*sd(data_1$HIGHEDC, na.rm=TRUE)/sd(data_1$conHipp_2h) #0.03737262 REVISION: 0.03816145 
fconHipp_2h["ThreatETA"]*sd(data_1$ThreatETA, na.rm=TRUE)/sd(data_1$conHipp_2h) #-0.03065245  

#Get predicted values
data_1$conHipp_2hyhat = predict(conHipp_2hFINAL, newdata = data_1, re.form = NA)

#Plot sig main effects of interest 
#First get +/- 2 SD for y axis 
mean(data_1$conHipp_2h) + 2*(sd(data_1$conHipp_2h)) #0.1814362
mean(data_1$conHipp_2h) - 2*(sd(data_1$conHipp_2h)) #-0.2149446

conHipp_2hlmerHIGHED <- ggplot(data = data_1, aes(x = HIGHED, y = conHipp_2hyhat)) +
  geom_jitter(aes(y=conHipp_2h),alpha = 0.1, size=.7) + 
  stat_smooth(method="lm", level = .95, lwd = 2, show.legend = FALSE) +
  theme_classic() +
  theme(axis.title.x = element_text(size=15),
        axis.title.y = element_text(size=15),
        axis.text.x = element_text (size=10),
        axis.text.y = element_text (size=10)) +
  scale_y_continuous(limits = c(-0.2149446, 0.1814362)) +
  ylab("Cingulo-Opercular and Hippocampus \n Connectivity") +
  xlab("Highest Parental Education (Years)")
conHipp_2hlmerHIGHED
ggsave("conHipp_2hlmerHIGHED.png", width = 5, height = 4)

conHipp_2hlmerTHREAT <- ggplot(data = data_1, aes(x = ThreatETA, y = conHipp_2hyhat)) +
  geom_jitter(aes(y=conHipp_2h),alpha = 0.1, size=.7) + 
  stat_smooth(method="lm", level = .95, lwd = 2, show.legend = FALSE) +
  theme_classic() +
  theme(axis.title.x = element_text(size=15),
        axis.title.y = element_text(size=15),
        axis.text.x = element_text (size=10),
        axis.text.y = element_text (size=10)) +
  scale_y_continuous(limits = c(-0.2149446, 0.1814362)) +
  ylab("Cingulo-Opercular and Hippocampus \n Connectivity") +
  xlab("Psychosocial Threat (MNLFA score)")
conHipp_2hlmerTHREAT
ggsave("conHipp_2hlmerTHREAT.png", width = 5, height = 4)

```





# conAcc_2h models
## Fit models

1. Fit with all random and fixed main effects including motion  
```{r conAcc_fit1, echo=TRUE, include=TRUE, cache=TRUE, message=FALSE}
conAcc_2hlmer1 <- lmer(paste('conAcc_2h', equation_1, sep = '~'), 
                    data = train, REML = FALSE,
                    control=lmerControl(optimizer = "Nelder_Mead"))
isSingular(conAcc_2hlmer1)
summary(conAcc_2hlmer1) 
```

No quadratic effects to test

3. Fit interactions  
```{r conAcc_2h_fit3, echo=TRUE, include=TRUE, cache=TRUE, message=FALSE}
#Income
conAcc_2hlmer3 <- lmer(paste('conAcc_2h', equation_1_plus_INRinteractions, sep = '~'), 
                     data = train, REML = FALSE,
                     control=lmerControl(optimizer = "Nelder_Mead"))
summary(conAcc_2hlmer3) #nothing, so moving on to edu now

# Edu
conAcc_2hlmer4 <- lmer(paste('conAcc_2h', equation_1_plus_EDinteractions, sep = '~'), 
                     data = train, REML = FALSE,
                     control=lmerControl(optimizer = "Nelder_Mead"))
summary(conAcc_2hlmer4) #none so going with model 1 REVISION: highedXsoceta sig

equation_1_plus_highedsoc = paste(equation_1, '+ HIGHEDC:SocETA')
conAcc_2hlmer5 <- lmer(paste('conAcc_2h', equation_1_plus_highedsoc, sep = '~'), 
                     data = train, REML = FALSE)
summary(conAcc_2hlmer5)

#compare
anova(conAcc_2hlmer1, conAcc_2hlmer5) #model 5 better


```
Final model = Model 5 (post revision).

## Test robustness with held-out data
```{r conAcc_2hTest, cache=TRUE, echo=TRUE}
R2 <- getR2(mod=conAcc_2hlmer5, outcome='conAcc_2h', test_dat=test, train_dat=train) 
R2 #0.0006212616 REVISION: 0.002281824
```

## Final model with complete dataset
```{r conAcc_2hFinal, warning=FALSE, echo=TRUE, include=TRUE, cache=TRUE}
conAcc_2hFINAL <- lmer(paste('conAcc_2h', equation_1_plus_highedsoc, sep = '~'), 
                    data = data_1, REML = FALSE,
                    control=lmerControl(optimizer = "Nelder_Mead"))
summary(conAcc_2hFINAL) #nothing REVISION: failed to converge despite some sig
```


# Save models 
```{r}

#First, double check that none have too high of VIF. A value above 5 indicates problematic co-linearity https://www.statology.org/variance-inflation-factor-r/

vif(dmnWithFINAL)
vif(salWithFINAL)
vif(danVanFINAL)
vif(conAmy_2hFINAL)
vif(conHipp_2hFINAL)
#Most values around 1, with depETA/INR around 3 suggesting moderate co-linearity for those vars but not problematic enough to remove


mods = list(dmnWithFINAL, salWithFINAL,
                   danVanFINAL,
                   conAmy_2hFINAL,conHipp_2hFINAL,
            conWithFINAL, danWithFINAL, vanWithFINAL, fparWithFINAL,
            salDmnFINAL, conAcc_2hFINAL
                   )

save(mods, file = paste(getwd(), '/final_mods.Rdata', sep = ''))

```
